Rename `classResults.py` -> `trades.py`.
